+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.2.0
+-----
+
+- Add serialization support for `Counter`;
+- Add union support for `Counter`;
+
 0.1.0 (initial release)
 -----------------------